format filters in scenario screening creation and update
