+++ packages/app-builder/src/repositories/ScreeningRepository.ts
@@ -61,52 +61,201 @@ export function makeGetScreeningRepository() {
       const listFeature: ScreeningAvailableFiltersAdapted =
         await marbleCoreApiClient.listScreeningAvailableFilters(feature);
       // FOR TESTING
-      // listFeature.sections.peps = {
-      //   topics: {
-      //     kind_of_peps: [{ name: 'primary', title: 'Primary PEP' }],
-
-      //     relevant_continents: [
-      //       { name: 'africa', title: 'Africa' },
-      //       { name: 'asia', title: 'Asia' },
-      //       { name: 'europe', title: 'Europe' },
-      //       { name: 'north-america', title: 'North America' },
-      //       { name: 'south-america', title: 'South America' },
-      //       { name: 'oceania', title: 'Oceania' },
-      //     ],
-      //     relevant_positions: [
-      //       { name: 'president', title: 'President' },
-      //       { name: 'prime-minister', title: 'Prime Minister' },
-      //       { name: 'minister', title: 'Minister' },
-      //       { name: 'senator', title: 'Senator' },
-      //       { name: 'representative', title: 'Representative' },
-      //       { name: 'member-of-parliament', title: 'Member of Parliament' },
-      //       { name: 'member-of-legislature', title: 'Member of Legislature' },
-      //       { name: 'member-of-congress', title: 'Member of Congress' },
-      //       { name: 'member-of-assembly', title: 'Member of Assembly' },
-      //       { name: 'member-of-senate', title: 'Member of Senate' },
-      //       { name: 'member-of-house', title: 'Member of House' },
-      //       { name: 'member-of-board', title: 'Member of Board' },
-      //       { name: 'member-of-committee', title: 'Member of Committee' },
-      //       { name: 'member-of-commission', title: 'Member of Commission' },
-      //       { name: 'member-of-council', title: 'Member of Council' },
-      //       { name: 'member-of-governing-body', title: 'Member of Governing Body' },
-      //       { name: 'member-of-governing-council', title: 'Member of Governing Council' },
-      //       { name: 'member-of-governing-board', title: 'Member of Governing Board' },
-      //       { name: 'member-of-governing-committee', title: 'Member of Governing Committee' },
-      //       { name: 'member-of-governing-commission', title: 'Member of Governing Commission' },
-      //       { name: 'ngo-member', title: 'NGO Member' },
-      //       { name: 'ngo-board-member', title: 'NGO Board Member' },
-      //       { name: 'ngo-committee-member', title: 'NGO Committee Member' },
-      //       { name: 'ngo-commission-member', title: 'NGO Commission Member' },
-      //       { name: 'ngo-council-member', title: 'NGO Council Member' },
-      //       { name: 'ngo-governing-body-member', title: 'NGO Governing Body Member' },
-      //       { name: 'ngo-governing-council-member', title: 'NGO Governing Council Member' },
-      //       { name: 'ngo-governing-board-member', title: 'NGO Governing Board Member' },
-      //       { name: 'ngo-governing-committee-member', title: 'NGO Governing Committee Member' },
-      //       { name: 'ngo-governing-commission-member', title: 'NGO Governing Commission Member' },
-      //     ],
+      // {
+      //   provider: 'lexisnexis',
+      //   sections: {
+      //     sanctions: {
+      //       self: 'sanctions',
+      //       datasets: [
+      //         { name: 'eulist', title: 'eulist' },
+      //         { name: 'li-fma-s', title: 'li-fma-s' },
+      //         { name: 'fr-minefe', title: 'fr-minefe' },
+      //         { name: 'mc-siccfin', title: 'mc-siccfin' },
+      //         { name: 'ofac', title: 'ofac' },
+      //         { name: 'ua-nsdc', title: 'ua-nsdc' },
+      //         { name: 'rosfinmon', title: 'rosfinmon' },
+      //         { name: 'seco', title: 'seco' },
+      //         { name: 'uk-fcolist', title: 'uk-fcolist' },
+      //         { name: 'au-dfat', title: 'au-dfat' },
+      //         { name: 'jp-mof', title: 'jp-mof' },
+      //         { name: 'ca-spmeru', title: 'ca-spmeru' },
+      //         { name: 'md-isc-s', title: 'md-isc-s' },
+      //         { name: 'au-cwlaw', title: 'au-cwlaw' },
+      //         { name: 'us-dos-s', title: 'us-dos-s' },
+      //         { name: 'nz-rsr', title: 'nz-rsr' },
+      //         { name: 'za-tfslist', title: 'za-tfslist' },
+      //         { name: 'ua-sfms', title: 'ua-sfms' },
+      //         { name: 'kr-mosf-s', title: 'kr-mosf-s' },
+      //         { name: 'va-lxviii', title: 'va-lxviii' },
+      //         { name: 'ar-repet-s', title: 'ar-repet-s' },
+      //         { name: 'il-modtl', title: 'il-modtl' },
+      //         { name: 'by-kgb-s', title: 'by-kgb-s' },
+      //         { name: 'hk-hkmas', title: 'hk-hkmas' },
+      //         { name: 'kr-kofiu', title: 'kr-kofiu' },
+      //         { name: 'nz-police', title: 'nz-police' },
+      //         { name: 'id-ppatk-s', title: 'id-ppatk-s' },
+      //         { name: 'tr-mtfa5', title: 'tr-mtfa5' },
+      //         { name: 'bh-mia-dtl', title: 'bh-mia-dtl' },
+      //         { name: 'us-ofacssi', title: 'us-ofacssi' },
+      //         { name: 'mx-uif1267', title: 'mx-uif1267' },
+      //         { name: 'un-sc-1267', title: 'un-sc-1267' },
+      //         { name: 'sa-pctc-nl', title: 'sa-pctc-nl' },
+      //         { name: 'ca-spmeir', title: 'ca-spmeir' },
+      //         { name: 'ca-spmebu', title: 'ca-spmebu' },
+      //         { name: 'il-mof', title: 'il-mof' },
+      //         { name: 'ae-ntlter', title: 'ae-ntlter' },
+      //         { name: 'us-dos-fto', title: 'us-dos-fto' },
+      //         { name: 'iq-tffc-nl', title: 'iq-tffc-nl' },
+      //         { name: 'mx-uif1988', title: 'mx-uif1988' },
+      //         { name: 'psca', title: 'psca' },
+      //         { name: 'tt-fiu-sl', title: 'tt-fiu-sl' },
+      //         { name: 'mx-uif1737', title: 'mx-uif1737' },
+      //         { name: 'un-sc-1988', title: 'un-sc-1988' },
+      //         { name: 'un-sc-1737', title: 'un-sc-1737' },
+      //         { name: 'uk-proter', title: 'uk-proter' },
+      //         { name: 'un-sc-1718', title: 'un-sc-1718' },
+      //         { name: 'tr-mtfa7', title: 'tr-mtfa7' },
+      //         { name: 'mx-uif1718', title: 'mx-uif1718' },
+      //         { name: 'pk-nacta-p', title: 'pk-nacta-p' },
+      //         { name: 'ph-atcsan', title: 'ph-atcsan' },
+      //         { name: 'id-ppa1718', title: 'id-ppa1718' },
+      //         { name: 'id-ppa1737', title: 'id-ppa1737' },
+      //         { name: 'qa-ntdl', title: 'qa-ntdl' },
+      //         { name: 'lv-fiu-s', title: 'lv-fiu-s' },
+      //         { name: 'us-doscuba', title: 'us-doscuba' },
+      //         { name: 'uz-dcec-nl', title: 'uz-dcec-nl' },
+      //         { name: 'kw-mofa-nl', title: 'kw-mofa-nl' },
+      //         { name: 'ca-spmeua', title: 'ca-spmeua' },
+      //         { name: 'au - ans', title: 'au - ans' },
+      //         { name: 'us-eo13382', title: 'us-eo13382' },
+      //         { name: 'tr-mtfa6', title: 'tr-mtfa6' },
+      //         { name: 'ca-spmeve', title: 'ca-spmeve' },
+      //         { name: 'kr-pmo', title: 'kr-pmo' },
+      //         { name: 'ng-nscnsl', title: 'ng-nscnsl' },
+      //         { name: 'pl-moia-s', title: 'pl-moia-s' },
+      //         { name: 'ca-spmeni', title: 'ca-spmeni' },
+      //         { name: 'pk-nacta-o', title: 'pk-nacta-o' },
+      //         { name: 'ca-syria', title: 'ca-syria' },
+      //         { name: 'my-moha-s', title: 'my-moha-s' },
+      //         { name: 'kz-fma-fte', title: 'kz-fma-fte' },
+      //         { name: 'nz-un-1373', title: 'nz-un-1373' },
+      //         { name: 'us-dos-tel', title: 'us-dos-tel' },
+      //         { name: 'in-mha', title: 'in-mha' },
+      //         { name: 'ch-fdfa-ve', title: 'ch-fdfa-ve' },
+      //         { name: 'ca-spmesd', title: 'ca-spmesd' },
+      //         { name: 'iq-tffc-il', title: 'iq-tffc-il' },
+      //         { name: 'ru-psd252', title: 'ru-psd252' },
+      //         { name: 'ru-fsb-tl', title: 'ru-fsb-tl' },
+      //         { name: 'un-sc-1591', title: 'un-sc-1591' },
+      //         { name: 'uk-hmt-ibt', title: 'uk-hmt-ibt' },
+      //         { name: 'tj - nbt', title: 'tj - nbt' },
+      //         { name: 'un-sc-2140', title: 'un-sc-2140' },
+      //         { name: 'kg-sfiu', title: 'kg-sfiu' },
+      //         { name: 'ca-jvcfoa', title: 'ca-jvcfoa' },
+      //         { name: 'ru-d1300', title: 'ru-d1300' },
+      //         { name: 'us-nsdnmbs', title: 'us-nsdnmbs' },
+      //         { name: 'tn-nsl', title: 'tn-nsl' },
+      //         { name: 'ca-riunrst', title: 'ca-riunrst' },
+      //         { name: 'fincen 311', title: 'fincen 311' },
+      //         { name: 'om-nctc-ll', title: 'om-nctc-ll' },
+      //         { name: 'kz-clssr', title: 'kz-clssr' },
+      //         { name: 'ca-spmeprc', title: 'ca-spmeprc' },
+      //         { name: 'ca-spmeby', title: 'ca-spmeby' },
+      //         { name: 'ke-pota-se', title: 'ke-pota-se' },
+      //         { name: 'sg-masfmr', title: 'sg-masfmr' },
+      //         { name: 'ca-spmehta', title: 'ca-spmehta' },
+      //         { name: 'un-sc-751', title: 'un-sc-751' },
+      //         { name: 'bis', title: 'bis' },
+      //         { name: 'us-ofaccmc', title: 'us-ofaccmc' },
+      //         { name: 'am-cba-dl', title: 'am-cba-dl' },
+      //         { name: 'ke-tfs-dl', title: 'ke-tfs-dl' },
+      //         { name: 'bg-com-s', title: 'bg-com-s' },
+      //         { name: 'lb-isf-ntl', title: 'lb-isf-ntl' },
+      //         { name: 'jo-tc-nl', title: 'jo-tc-nl' },
+      //         { name: 'ca-spmemd', title: 'ca-spmemd' },
+      //         { name: 'ch-fdfa-sy', title: 'ch-fdfa-sy' },
+      //         { name: 'lt-fcis-s', title: 'lt-fcis-s' },
+      //         { name: 'cz-mfa-nsl', title: 'cz-mfa-nsl' },
+      //         { name: 'us-eo12938', title: 'us-eo12938' },
+      //         { name: 'bd-bfiu-ds', title: 'bd-bfiu-ds' },
+      //         { name: 'ca-spmelk', title: 'ca-spmelk' },
+      //         { name: 'uk-rusr18a', title: 'uk-rusr18a' },
+      //         { name: 'bf-ccga-nl', title: 'bf-ccga-nl' },
+      //         { name: 'ca-trcl', title: 'ca-trcl' },
+      //         { name: 'us-cbws', title: 'us-cbws' },
+      //         { name: 'pl-gifi-s', title: 'pl-gifi-s' },
+      //         { name: 'au-ans-sst', title: 'au-ans-sst' },
+      //         { name: 'cz-ism-ct', title: 'cz-ism-ct' },
+      //         { name: 'sg-trcl', title: 'sg-trcl' },
+      //         { name: 'us-eo13808', title: 'us-eo13808' },
+      //         { name: 'un-sc-2653', title: 'un-sc-2653' },
+      //         { name: 'us-nppa', title: 'us-nppa' },
+      //         { name: 'at-onb', title: 'at-onb' },
+      //         { name: 'ca-spmegt', title: 'ca-spmegt' },
+      //         { name: 'eg-dtl', title: 'eg-dtl' },
+      //         { name: 'sg-agc', title: 'sg-agc' },
+      //         { name: 'un-1718-dv', title: 'un-1718-dv' },
+      //         { name: 'au-ans-phg', title: 'au-ans-phg' },
+      //         { name: 'ca-spmeht', title: 'ca-spmeht' },
+      //         { name: 'cn-taosc', title: 'cn-taosc' },
+      //         { name: 'mz-gifim-s', title: 'mz-gifim-s' },
+      //         { name: 'un-sc-1533', title: 'un-sc-1533' },
+      //         { name: 'us-ds-sst', title: 'us-ds-sst' },
+      //         { name: 'us-eo14064', title: 'us-eo14064' },
+      //         { name: 'kz-fma-ita', title: 'kz-fma-ita' },
+      //         { name: 'lk-fiu1373', title: 'lk-fiu1373' },
+      //         { name: 'ma-cnasnu', title: 'ma-cnasnu' },
+      //         { name: 'tw-mjib-sl', title: 'tw-mjib-sl' },
+      //         { name: 'uk-dbt-cl', title: 'uk-dbt-cl' },
+      //         { name: 'us-bis-cl', title: 'us-bis-cl' },
+      //         { name: 'us-dos231a', title: 'us-dos231a' },
+      //         { name: 'cn-mfasana', title: 'cn-mfasana' },
+      //         { name: 'eu-trcl', title: 'eu-trcl' },
+      //         { name: 'my-miti-cl', title: 'my-miti-cl' },
+      //         { name: 'us-eiba', title: 'us-eiba' },
+      //         { name: 'au-dfat-cl', title: 'au-dfat-cl' },
+      //         { name: 'cn-mps-s', title: 'cn-mps-s' },
+      //         { name: 'rw-fic-dsl', title: 'rw-fic-dsl' },
+      //         { name: 'so-namlc-s', title: 'so-namlc-s' },
+      //         { name: 'un-sc-2048', title: 'un-sc-2048' },
+      //         { name: 'un-sc-2127', title: 'un-sc-2127' },
+      //         { name: 'un-srcl', title: 'un-srcl' },
+      //         { name: 'ch-seco-te', title: 'ch-seco-te' },
+      //       ],
+      //     },
+      //     peps: {
+      //       self: 'pep',
+      //       topics: {
+      //         geography: [
+      //           { name: 'pep.geo.eu', title: 'pep.geo.eu' },
+      //           { name: 'pep.geo.us', title: 'pep.geo.us' },
+      //         ],
+      //         kind: [
+      //           { name: 'pep.kind.primary', title: 'pep.kind.primary' },
+      //           { name: 'pep.kind.secondary', title: 'pep.kind.secondary' },
+      //         ],
+      //         position: [
+      //           { name: 'pep.position.headofstate', title: 'pep.position.headofstate' },
+      //           { name: 'pep.position.legislative', title: 'pep.position.legislative' },
+      //         ],
+      //         status: [
+      //           { name: 'pep.status.active', title: 'pep.status.active' },
+      //           { name: 'pep.status.inactive', title: 'pep.status.inactive' },
+      //         ],
+      //       },
+      //     },
+      //     adverse_media: {
+      //       self: 'adversemedia',
+      //       topics: {
+      //         source: [
+      //           { name: 'adversemedia.media', title: 'adversemedia.media' },
+      //           { name: 'adversemedia.enforcements', title: 'adversemedia.enforcements' },
+      //         ],
+      //       },
+      //     },
       //   },
       // };
+
       // listFeature.conditional_filters = [
       //   {
       //     key: 'kind_of_peps',
